Currently refactoring.
I've added a new method getAccountType that now handles the while loop which asks the user to input the account type.

I've also added a method called askQuestion, which is an alternate way of prompting the user for their personal details when creating a new account.

From the testing I've done everything seems to work just as efficiently as it had before.