GetAccountType removed & list implemented for account types instead.
I'm going to attempt to take things in a different direction. I may revert back to what I was doing earlier if this doesn't work out.

The output is fine and everything seems to be working, although the word "savings" is printed twice in the output in the create account section and I have no idea why.