Refactored the code more so that there's less duplication/repitition of code throughout.
ie. created a method which handles the amount a user wants to withdraw or deposit, this handles depositCash() and withdrawCash(). Prior to this they had more or less the same code inside each of them.

Also created a new class called InvalidAccountTypeException, which simply throws an error if an invalid account is created.