Made a small change...
The phrase "Please choose an account based on the number assigned . . . " repeated in two places where it shouldn't have because I put it in the wrong place initially, and forgot to get rid of it. Also it needed to be inside an else statement so that it wouldn't be printed before every account pops up. Now it only is printed once when necessary.